refactor(backend): moving function to check inabilities in a channel to the membership service
